analyze/requirements.txt
Pillow~=9.4.0
requests~=2.28.2
setuptools~=58.1.0
numpy~=1.23.5
matplotlib~=3.7.1
torch~=2.0.1
LAC~=2.1.2
docx2txt~=0.8
python-dateutil~=2.8.2
docx~=0.2.4
pypdf2~=3.0.1
docx2pdf~=0.1.8
pypdf==3.8.1